data: remove the entry for the Gaomon S56K
Same USB ID and name as the Huion Kamvas Pro 13 added in
43acb1f7abe8c3a9595c708918ae6e9f0e1e51a1 but since the Kamvas Pro has more
information and needs the SVG plus button count, let's drop this file here.

This means the Gaomon will show up as Kamvas now with buttons that can't be
mapped but until someone convinces HUION to not re-use USB IDs that's
unavoidable.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>