CI: don't enable -Wall -Wextra in the CI
This is a leftover from autotools builds, meson uses the warning_level
option for this. The net result of our CFLAGS being appended was this
sequence:
  -Wall -Wextra -Wno-foo -Wall -Wextra

The second set of -Wall -Wextra undid all the -Wno-foo we carefully
selected in meson.build - at least with clang, possibly with gcc too.

This means we can now drop -Wno-error=unused-parameter because that's
already disabled in meson.build.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>