Add files via upload
I've done some basic testing and it works, but for some reason it displays the tablet display twice in gnome-control-center.
I'm not sure what to check to further troubleshoot, but this does allow me to configure the tablet and pen in gnome-control-center, so I figured I'd submit a pull request first.  

This used kamvas-pro-13.tablet as a basis and I just copied/renamed the kamvas-pro-13.svg since the button layout is functionally the same.

Very new to contributing, etc., so apologies for any missteps.