data: add a new Keys/KeyCodes entry to the tablet files
This is parsed but not exported by libwacom itself beyond the number of
available keys. We can add other bits once we know what potential
callers want.

It is used however to determine if a device has a pad and thus used to
set the ID_INPUT_TABLET_PAD property through the hwdb. This fixes the
issue with some devices without buttons (but with keys) to not get
labelled as pads, e.g. the Cintiq Pro 13.

Fixes #585