test: add some tracing to the svg validity test
It's a bit hard to figure out from the CI logs what failed because they
may be truncated. And for users not intimitely familiar with the test
suite it's not that simple either so let's put in a bunch of trace
stataments that should at least point to where things go wrong.