Use NumRings instead of separate Ring/Ring2 (#669)
This deprecates libwacom_has_ring() and libwacom_has_ring2() for the
new libwacom_get_num_rings(). In our tablet files we also remove
Ring=true/Ring2=true and replace that with NumRings=.

This makes the support line up better with strip support and removes
some of the confusion in the tablet file where we have two independent
Ring and Ring2 entries with different meanings.