Rework the match approach to prefer uniq over name
Previously we would match for name first, then uniq. But the name
matches are less reliable than uniq matches so where we can match for
uniq we should prefer that.

This avoids the case where where a name that applies to multiple
devices takes precedence over a more precise uniq match.