test: replace the uinput-based hwdb test with a direct query
Previously we'd create a uinput device for each of our tablets, then
query udev if the property is set. This takes a long time, in parts
thanks to the 300ms sleep to wait for the uinput device to be ready.

With now over 500 tablet files and most of those having multiple matches
we're looking at at minimum 500*0.3 seconds for a test run which isn't
really useful for a test like this.

Instead - let's query systemd-hwdb directly: construct a query string
our udev rule and query systemd-hwdb for the properties, then ensure we
have the ones we expect.

Closes #675