layouts: Fix buttons for Huion and Gaomon tablets (#691)
libwacom's SVG format requires that we have an element with
id="ButtonA", one with id="LabelA" and one with id="LeaderA".

The latter is the line pointing to the button. Since OSDs mostly fill in
the button rect when the button is pressed we need the label to be
outside the button and a leader pointing to it.

This patch was (supervised but) generated from a Python script:
all labels are moved to the right and a leader with the right id and
class was inserted. The script supports moving buttons (and adding
leaders) in all four directions, most tablets are ok now. Some tablets
have a less-than optimal layout (e.g. the Huion Mini Keypad) but that
can be fixed by someone with more time and motivation.

Script source:
https://gist.github.com/whot/8f3dbc09908862775b61f7d0a71fd5ed

Fixes: c80bd7f7bfdf ("Add auto generated GAOMON tablet description files")
Fixes: ce093e730fb1 ("Add auto generated HUION tablet description files")