data: drop the firmware match from the Huion Inspiroy 2 S
This devices appears to have its own unique PID so we don't need the
firmware match here. This matches the M and L version which also have
a unique PID.

If Huion or Gaomon end up releasing devices that share this PID we'll
have to re-instated the firmware match but for now let's drop it.