Support styli from vendors other than Wacom
The current API uses tool IDs that are bound to Wacom devices
(currently only wacom devices use those). If we have styli from other
vendors that use the same tool ID we cannot currently express this
without having duplicates.

This patch introduces an internal WacomStylusId struct with vid and
tool_id and switches the internal code over to use that.
It deprecates the current stylus id functions and replaces them with
a set that returns the WacomStylus pointer to the caller instead of just
the id. Where the id is actually used, the caller can then query that
struct.

A special case are the @generic styli which are assigned a VID of zero.

Internal behaviour treats Wacom styli still special: styli with the
0x56a vendor id will not have a vid prefix but just the default tool id.
Likewise, the legacy functions libwacom_get_supported_styli() only returns
Wacom and the generic styli.