tools/clean_svg: allow passing in a .tablet file
If the svg file passed in is a .tablet file, extract the layout and
tablet name from there. This allows us to run the tool on the existing
tablet files to check for any differences.

Add this into the CI so we ensure that script doesn't crash on our SVG
files.