Generate the device class if not explicitly given in the tablet file
libwacom_get_class() is deprecated but it's still a part of the API, and
thus have required even new .tablet files to fill this out even though
it's superseded by IntegratedIn. So let's generate them programmatically.

This turns devices that are IntegratedIn=Display to the Cintiq class,
IntegratedIn=Display;System into the ISDV4 class, and everything else
into the Bamboo class (because Bamboo is not expected to have tool
ids/serials).