Generate the class given in class if not explicitly given
This is still a part of the API, and thus have required even new .tablet
files to fill this out even though it's superseded by IntegratedIn. So
let's generate them programmatically.

This turns devices that are IntegratedIn=Display to the Cintiq class,
IntegratedIn=Display;System into the ISDV4 class, and everything else
into the Bamboo class.