Change fallback return value for libwacom_get_button_led_group to an int
The implementation looked like this returns an enum value but it
doesn't, the returned value is the numerical index of the LED group
(typically 0 and 1 for Ring/Ring2 but also 0 and 1 for Strip/Strip2
since we don't have devices with combined features).

Let's change the fallback to -1 too instead of the enum - same value but
it makes the code more obvious.