data: drop the Class assignment for Huion, Gaomon and XP Pen devices
The class is deprecated anyway, no-one appears to be using this for
anything but EK Remote detection and the classes make no sense for any
non-Wacom devices anyway. Let's remove the explicit classes and
generate them from the integration flags instead - this way at least we
have some consistent behavior.