Add support for getting the target mode for a modeswitch button
Add libwacom_get_button_modeswitch_mode() to return the target
mode for a given button. On the vast majority of devices this
will be "next" but the Wacom Mobile Studio and the Cintiq 24HD have
multiple mode switch buttons that correspond to the respective mode
(i.e. button N should switch to mode M).

This is already effectively encoded in our tablet database so we
can simply use the order of buttons in that file to assign the modes.

Closes #854