layouts: rearrange the Intuos Pro 3rd L/M buttons
The current layout is too far on the left/right edge, causing the right
set of buttons to be nonconfigurable in GNOME's OSD, see e.g.
https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/3500

Re-do the layouts so the button configuration is below the dials.
Ideally the GNOME OSD is smart enough so this isn't an issue to start
with but since we have loads of empty space here anyway, let's
make this a simple fix here until the rest can catch up.