data: add unmangled stylus IDs to wacom.stylus
Years ago, an extra '0' accidently landed on the 4th nybble of the
stylus IDs in Wacom kernel driver. We decided to fix the issue in
the kernel (https://github.com/linuxwacom/libwacom/issues/767).
But, we	want libwacom to be prepared for the change.

The "Mangled tool ID for kernel < 6.20"	comments in the	patch is
only a placeholder. We will replace 6.20 with the actual kernel
version that the mangled IDs are replaced by the correct ones.